Move LoginRequiredTestMixin to the core app
